… ggml/src/ggml-qnn/npu/device/op_impl.cpp
… ggml/src/ggml-qnn/npu/device/op_mul_mat.cpp
… ggml/src/ggml-qnn/npu/device/op_rope.cpp
… ggml/src/ggml-qnn/npu/device/tensor.hpp
… ggml/src/ggml-qnn/npu/device/thread_pool.hpp
… ggml/src/ggml-qnn/npu/device/type_traits.cpp
… ggml/src/ggml-qnn/npu/device/type_traits.hpp
… ggml/src/ggml-qnn/npu/device/util.hpp
… ggml/src/ggml-qnn/npu/device/vec_ops.cpp
… ggml/src/ggml-qnn/npu/device/vec_ops.hpp
… ggml/src/ggml-qnn/npu/device/vec_ops.inl
… ggml/src/ggml-qnn/npu/host/buffer.cpp
… ggml/src/ggml-qnn/npu/host/graph.cpp
… ggml/src/ggml-qnn/npu/host/graph.hpp
… ggml/src/ggml-qnn/npu/host/tensor.hpp
… ggml/src/ggml-qnn/npu/idl/hexagon_npu.idl